refactor(balance): Move balance to Movements namespace
