+++ lib/fintoc/resources/movements/balance.rb
@@ -1,0 +1,27 @@
+# frozen_string_literal: true
+
+module Fintoc
+  module Movements
+    class Balance
+      attr_reader :available, :current, :limit
+
+      def initialize(available:, current:, limit:)
+        @available = available
+        @current = current
+        @limit = limit
+      end
+
+      def id
+        object_id
+      end
+
+      def to_s
+        "#{@available} (#{@current})"
+      end
+
+      def inspect
+        "<Fintoc::Movements::Balance #{@available} (#{@current})>"
+      end
+    end
+  end
+end
